Set up the Product module with all required layers including DTOs

+++ src/main/java/com/ecommerce/Ecommerce_App/Model/Product.java
@@ -22,9 +22,9 @@ public class Product {
     @NotBlank
     private String productDescription;
     @NotNull
-    private Integer price;
+    private double price;
     @NotNull
-    private Integer specialPrice;
+    private double specialPrice;
     @NotNull
     private Integer quantity;
